pyproject.toml
[project]
name = "prakt021"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
]
